Refactor .Position to support pawn-originated captures.
Moving vectors are now a sumtype which has a constructor that lets us
determine whether or not the move is a "nomral" move or one caused by a
pawn capturing diagonally. To make use of this, we have to carry around
some extra information in the machinery in the Position module.

Signed-off-by: Ricky Elrod <ricky@elrod.me>