Revert "Disallow pawn-moves-forward capture"
This reverts commit 2a37fa27fbe16012d4158ebe6ee649cd7dacd234.

This commit is wrong, it checks the piece at the target, not the moving
piece. We will have to carry around a bit more information such as the
starting square (so we can index into the board and see what kind of
piece it is), or information about the piece directly.